Change activation function for log r
The activation function for log r was the rectify function, thus
excluding values less than 1 for r. Similarly, the clipping in the
function to calculate the logarithm of the negative binomial
distribution was also change to allow for values below 1 for r.

Also, p and log r are saved for the reconstructions, and the metadata
and the number of epochs trained is saved for the model.