Change loading of clusters
Clusters were loaded as dictionary of cell IDs with corresponding
cluster ID. It proved easier to load them as a dictionary of cluster IDs
with a corresponding list of cell IDs.

Colours used for each cluster used in Macosko et al. (2015) were also
transcribed to the data module, and the main module as well as the run
script were changed to read in the cluster data file.